+++ src/compojure/api/routes.clj
@@ -8,6 +8,7 @@
             [ring.swagger.common :as rsc]
             [clojure.string :as str]
             [linked.core :as linked]
+            [compojure.response]
             [schema.core :as s])
   (:import [clojure.lang AFn IFn Var]))
 
@@ -41,7 +42,10 @@
   ([handler]
    (get-routes handler nil))
   ([handler options]
-   (-get-routes handler options)))
+   (mapv
+     (fn [route]
+       (update-in route [0] (fn [uri] (if (str/blank? uri) "/" uri))))
+     (-get-routes handler options))))
 
 (defrecord Route [path method info childs handler]
   Routing
@@ -49,9 +53,21 @@
     (let [valid-childs (filter-routes this options)]
       (if (seq childs)
         (vec
-          (for [[p m i] (mapcat #(get-routes % options) valid-childs)]
+          (for [[p m i] (mapcat #(-get-routes % options) valid-childs)]
             [(->paths path p) m (rsc/deep-merge info i)]))
         (into [] (if path [[path method info]])))))
+
+  compojure.response/Renderable
+  (render [_ {:keys [uri request-method]}]
+    (throw
+      (ex-info
+        (str "\ncompojure.api.routes/Route can't be returned from endpoint "
+             (-> request-method name str/upper-case) " \"" uri "\". "
+             "For nested routes, use `context` instead: (context \"path\" []  ...)\n")
+        {:request-method request-method
+         :path path
+         :method method
+         :uri uri})))
 
   IFn
   (invoke [_ request]